backend/requirements.txt
fastapi
uvicorn

google-generativeai
python-dotenv
pydantic
requests
